Final small edits

+++ RandomWiki/Core/Components/UI/CustomSlider.swift
@@ -13,7 +13,8 @@ struct CustomSlider: UIViewRepresentable {
     var min: Float
     var max: Float
     var step: Float
-    @Binding var value: Float
+    
+    @Binding var value: Double
     
     func makeUIView(context: Context) -> UISlider {
         let slider = UISlider(frame: .zero)
@@ -22,14 +23,14 @@ struct CustomSlider: UIViewRepresentable {
         slider.maximumTrackTintColor = .gray
         slider.minimumValue = self.min
         slider.maximumValue = self.max
-        slider.value = value
+        slider.value = Float(value)
         slider.isContinuous = true
         slider.addTarget(context.coordinator, action: #selector(Coordinator.update(_:)), for: .valueChanged)
         return slider
     }
     
     func updateUIView(_ uiView: UISlider, context: Context) {
-        uiView.value = self.value
+        uiView.value = Float(self.value)
     }
     
     func makeCoordinator() -> CustomSlider.Coordinator {
@@ -37,15 +38,15 @@ struct CustomSlider: UIViewRepresentable {
     }
     
     final class Coordinator: NSObject {
-        var value: Binding<Float>
+        var value: Binding<Double>
         var step: Float
-        init(value: Binding<Float>, step: Float) {
+        init(value: Binding<Double>, step: Float) {
             self.value = value
             self.step = step
         }
         @objc func update(_ sender: UISlider) {
             let steppedVal = roundf(sender.value / step) * step
-            self.value.wrappedValue = steppedVal
+            self.value.wrappedValue = Double(steppedVal)
             sender.value = steppedVal
         }
     }
